Fix: SafeApprove function not available in IERC20 interface
The main change is to use the standard approve function instead of safeApprove in the sendHashWithToken function. We also added a require statement to ensure that the approve operation was successful.

This change should resolve the previous error message. The approve function is part of the ERC20 standard and must be implemented by all ERC20 tokens.