change WormholeMessenger
The IWormhole interface is defined within the contract. This provides an implementation closer to Wormhole's real interface and avoids possible compilation errors.
VM structure and related constants (VM_STATUS_VALID, VM_STATUS_INVALID) have been added to the IWormhole interface.
The signature of the publishMessage function has been made more compatible with the actual implementation of Wormhole.
In the sendHashWithToken function, standard approve was used instead of safeApprove and a require control was added.
All functions and variables are defined with consistent and correct types.